boo intercept
Turns out the intercept does bad things. Have gotten rid of it and it matches with Max's old one for simple cases. It gives better likelihood estimates as well. One worrisome aspect is the fact that none of these ever get the tau terms right, and it seems to throw things off.