Change tau's
Worked on more endogenous regressors. The simulation code can't handle it, but it looks like the wrapper can, for now. Also discovered that the way I've been doing the log-likelihood bit won't work. In the name pieces function I input some extra info. Because of the way the optimization code works in R, I think it will try to optimize that info, which will throw everything off. So that whole function has to be redone.