Changed make.covTrans
Made a new version of the function to make and transform the covariance matrix. The new version should replace the one that used to live in hurdleIV.sim. It should be able to take in more kinds of inputs (the full matrix, the matrix as a vector, or a list of the useful parameters).