Redo Changing gamma tobit bias
This file creates the figure for the bias in Tobit estimates when gamma and beta start to diverge. It isn't as dramatic as before (I think I was generating the data with the wrong errors). I am still suspicious however. When gamma is at it's assumed value, beta should be its "true" value I think. This worked in the old code where I didn't use the A matrix transformation of the errors. It may be I need to go back and redo the calculations to tell me what the value "should" be.